add showViewContainer setting & refactor code
