+++ src/constants/configs.ts
@@ -1,3 +1,6 @@
+import { StatusBarAlignment } from "vscode";
+
+/* eslint-disable @typescript-eslint/naming-convention */
 const FAVOURITES_SAVING_PROFILE = {
   NAME: {
     AUTO: "auto",
@@ -7,6 +10,12 @@ const FAVOURITES_SAVING_PROFILE = {
   DEFAULT: "global",
 };
 
+const STATUS_BAR_ITEM = {
+  ALIGNMENT: StatusBarAlignment.Right,
+  PRIORITY: 100,
+};
+
 export default {
   FAVOURITES_SAVING_PROFILE,
+  STATUS_BAR_ITEM,
 };
